pubsub / events: centralize publication code
